move TopicUrlDTO to base
