+++ src/main/java/mServer/crawler/sender/srf/tasks/SrfTopicsOverviewTask.java
@@ -10,7 +10,7 @@ import mServer.crawler.sender.MediathekReader;
 import mServer.crawler.sender.ard.tasks.ArdTaskBase;
 import mServer.crawler.sender.base.AbstractRecursivConverterTask;
 import mServer.crawler.sender.base.CrawlerUrlDTO;
-import mServer.crawler.sender.orf.TopicUrlDTO;
+import mServer.crawler.sender.base.TopicUrlDTO;
 import mServer.crawler.sender.srf.SrfConstants;
 import mServer.crawler.sender.srf.parser.SrfTopicsDeserializer;
